refactor: move Phel proxy static methods out from original
